+++ ChangeLog.md
@@ -6,5 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- imgproc.matchHist(): "Modify image intensities to match the histogram of a reference image" by rcasero
-- testdata/*.png: Stereo cloud images with ROI masks (left_mask.png, left.png, right_mask.png, right.png) by rcasero
+
+- imgproc.matchHist(): "Modify image intensities to match the
+  histogram of a reference image" by rcasero
+
+- imgproc.imfuse(): "Composite of two images" by rcasero
+
+- testdata/*.png: Stereo cloud images with ROI masks (left_mask.png,
+  left.png, right_mask.png, right.png) by rcasero